requirements.txt
# Core Dependencies
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Document Processing
PyPDF2==3.0.1
pdf2image==1.16.3
python-docx==1.1.0
Pillow==10.1.0
pytesseract==0.3.10

# NLP and ML
transformers==4.36.0
torch==2.1.0
spacy==3.7.2
scikit-learn==1.3.2
numpy==1.26.2
pandas==2.1.4

# Clinical NLP
medspacy==0.2.0.0
scispacy==0.5.3

# API and Web
fastapi==0.108.0
uvicorn==0.25.0
pydantic==2.5.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.23
alembic==1.13.1

# Anonymisation
presidio-analyzer==2.2.33
presidio-anonymizer==2.2.33

# Utilities
python-dateutil==2.8.2
regex==2023.12.25
cryptography==41.0.7

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0

# Logging and Monitoring
loguru==0.7.2
